+++ src/main/java/com/codetaylor/mc/pyrotech/modules/core/command/ClientCommandLang.java
@@ -1,15 +1,18 @@
 package com.codetaylor.mc.pyrotech.modules.core.command;
 
-import com.codetaylor.mc.pyrotech.modules.core.ModuleCore;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.entity.EntityPlayerSP;
 import net.minecraft.command.CommandBase;
 import net.minecraft.command.CommandException;
 import net.minecraft.command.ICommandSender;
+import net.minecraft.item.ItemStack;
 import net.minecraft.server.MinecraftServer;
+import net.minecraft.util.text.TextComponentTranslation;
 import net.minecraftforge.client.IClientCommand;
 
 import javax.annotation.Nonnull;
 
-public class ClientCommandExport
+public class ClientCommandLang
     extends CommandBase
     implements IClientCommand {
 
@@ -23,31 +26,27 @@ public class ClientCommandExport
   @Override
   public String getName() {
 
-    return "ptexport";
+    return "ptlang";
   }
 
   @Nonnull
   @Override
   public String getUsage(@Nonnull ICommandSender sender) {
 
-    return "ptexport";
+    return "ptlang";
   }
 
   @Override
   public void execute(@Nonnull MinecraftServer server, @Nonnull ICommandSender sender, @Nonnull String[] args) throws CommandException {
 
-    int size = 512;
+    Minecraft minecraft = Minecraft.getMinecraft();
+    EntityPlayerSP player = minecraft.player;
 
-    if (args.length > 0) {
+    ItemStack heldItemMainhand = player.getHeldItemMainhand();
 
-      try {
-        size = Integer.valueOf(args[0]);
-
-      } catch (Exception e) {
-        throw new CommandException("", e);
-      }
+    if (!heldItemMainhand.isEmpty()) {
+      String unlocalizedName = heldItemMainhand.getUnlocalizedName();
+      sender.sendMessage(new TextComponentTranslation(unlocalizedName));
     }
-
-    ModuleCore.Utils.BLOCK_RENDERER.render(size);
   }
 }